DistanceMatrix refactor Phase D+E: fix remaining index-related narrowing sites, final sign-off
Phase D, scoped by direct discussion to only the DistanceMatrix-
adjacent size_t->int findings (most of the 150 remaining after Phase C
turned out to be an unrelated problem - double->float precision loss
in formula code, stream/char sentinel handling, libxml2 API calls -
left alone, confirmed unrelated by reading each site rather than
assumed from its finding category):

- compute_Tamura_Nei/compute_Tamura_Nei_fixratio's must-stay-int
  strlen/numAs/numCs/numGs/numTs params (lint_plan.md's Phase 2 already
  established why - a signed subtraction inside them) - cast at all 11
  call sites this refactor's own Phase C code introduced.
- NeighborJoining.cpp's computeNJTree(): found the same must-stay-
  signed constraint in a new place - counts i down through 0 with
  i >= 0, so widening rows/columns instead of casting would have been
  a real, live infinite-loop/out-of-bounds bug, not just a lint
  regression. Cast with a comment explaining why.
- str2int_hashmap's int value type (LeastSquaresFit.cpp, fnj/main.cpp)
  - cast at the two call sites rather than changing a project-wide
  shared type.
- DataOutputStream::printRow()'s virtual hierarchy (base + all three
  overrides, widened together so the override set stays matched) and
  each override's entriesPerRow local - genuinely safe to widen, all
  simple upward-counting loops with no signed dependency.
  PhylipDmOutputStream.cpp had one site doing a pointless
  int-then-back-to-size_t double-cast; widening removed both casts,
  not just silenced the warning.
- applyFixFactor/applyFixFactorRow (DistanceMatrix/FloatDistanceMatrix/
  DistanceRow.cpp) - their own free functions, same treatment.

Measured via clang-tidy narrowing-conversions diff (regenerated
compile_commands.json after one run hit a spurious tooling error on
TamuraNei.cpp - a stale include-path cache, not a real issue, since
that file was never touched here and the real build succeeded
throughout): 150 -> 111 findings, all 39 resolved sites confirmed
DistanceMatrix-adjacent, every remaining one re-confirmed unrelated.

Phase E: full clean rebuild (0 errors), ctest (2/2), RunExamples.sh
(byte-identical), RunCliChecks.sh (61/61), bench_primitives (healthy
speedups, no regression), plus a DNA-distance sanity check across all
four models on a 300-seq/8000bp synthetic dataset.

All 5 phases of distance_matrix_refactor_plan.md are now complete.

+++ include/fastphylo/io/PhylipDmOutputStream.hpp
@@ -24,7 +24,7 @@ public:
   // rewrite; still shared here rather than duplicated, since the
   // ONEDIGIT/TENDIGIT-based formatting it uses is identical to
   // printPHYLIPfastSD's.
-  void printRow(StrFloRow &dm, std::string name, int row, bool mem_eff_flag) override;
+  void printRow(StrFloRow &dm, std::string name, size_t row, bool mem_eff_flag) override;
   // fastdist's main.cpp calls printHeader() unconditionally before its
   // printRow() streaming loop, needing it to actually write the count
   // line (there's no other point at which that line would be written in
