+++ src/fastphylo/core/DistanceRow.cpp
@@ -13,9 +13,9 @@ using std::endl;
 bool
 applyFixFactorRow(StrFloRow &dm, float fixFactor){
   float biggest = 0;
-  int size =dm.getColumns();
+  size_t size =dm.getColumns();
 
-	for ( int j = 0 ; j < size ; j++ ){
+	for ( size_t j = 0 ; j < size ; j++ ){
 	  float d = dm.getDistance(j);
 	  if ( isfinite(d) && d>biggest ){
 	    biggest = d;
@@ -24,7 +24,7 @@ applyFixFactorRow(StrFloRow &dm, float fixFactor){
 
   bool changed = false;
   biggest = biggest*fixFactor;
-	for ( int j = 0 ; j < size ; j++ ){
+	for ( size_t j = 0 ; j < size ; j++ ){
 	  float d = dm.getDistance(j);
 	  if ( !isfinite(d) || d<0 ){
 	    dm.setDistance(j,biggest);
